refactor: separate preference state policy
