+++ ui/src/app/server-prefs.ts
@@ -4,114 +4,23 @@
 // intent shadows server snapshots until the hash-free LWW ack; failed pushes degrade device-local.
 import { asNullableRecord as asRecord } from "@openclaw/normalization-core/record-coerce";
 import type { GatewayBrowserClient } from "../api/gateway.ts";
-import { normalizeSidebarEntries } from "../app-navigation.ts";
-import { isSupportedLocale } from "../i18n/index.ts";
 import type { RuntimeConfigCapability } from "../lib/config/index.ts";
 import {
-  loadSettings,
-  normalizeChatFollowUpModeOverride,
-  normalizeChatSendShortcut,
-  patchSettings,
-  UI_APPEARANCE_DEFAULTS,
-  type ChatFollowUpMode,
-  type ChatSendShortcut,
-  type UiSettings,
-} from "./settings.ts";
-import type { ThemeMode, ThemeName } from "./theme.ts";
-const THEMES: ReadonlySet<ThemeName> = new Set(["claw", "knot", "dash", "custom"]);
-const THEME_MODES: ReadonlySet<ThemeMode> = new Set(["light", "dark", "system"]);
-type SyncedPrefSpec<T> = {
-  extract: (value: unknown) => T | undefined;
-  local: (settings: UiSettings) => T | undefined;
-  write?: (value: T | undefined) => Partial<UiSettings>;
-  canApply?: (value: T, settings: UiSettings) => boolean;
-  clearable?: boolean;
-  reset?: (settings: UiSettings) => Partial<UiSettings>;
-};
-const prefSpec = <T>(specification: SyncedPrefSpec<T>) => specification;
-function prefValuesEqual(left: unknown, right: unknown): boolean {
-  if (Array.isArray(left) && Array.isArray(right)) {
-    return left.length === right.length && left.every((value, index) => value === right[index]);
-  }
-  return left === right;
-}
-/**
- * One descriptor per synced pref — the single source of truth for what syncs
- * through config ui.prefs. Each key defines server validation, local normalization,
- * and applicability; `clearable` keys push an explicit JSON null when unset locally
- * so the merge patch removes them server-side.
- */
-const SYNCED_PREFS = {
-  theme: prefSpec<ThemeName>({
-    extract: (value) => (THEMES.has(value as ThemeName) ? (value as ThemeName) : undefined),
-    local: (settings) => settings.theme,
-    write: (value) => ({ theme: value ?? UI_APPEARANCE_DEFAULTS.theme }),
-    clearable: true,
-    reset: () => ({ theme: UI_APPEARANCE_DEFAULTS.theme }),
-    // A server "custom" theme is only honorable once this browser imported one;
-    // the imported palette itself is too large to live in config.
-    canApply: (value, settings) => value !== "custom" || Boolean(settings.customTheme),
-  }),
-  themeMode: prefSpec<ThemeMode>({
-    extract: (value) => (THEME_MODES.has(value as ThemeMode) ? (value as ThemeMode) : undefined),
-    local: (settings) => settings.themeMode,
-    write: (value) => ({ themeMode: value ?? UI_APPEARANCE_DEFAULTS.themeMode }),
-    clearable: true,
-    reset: () => ({ themeMode: UI_APPEARANCE_DEFAULTS.themeMode }),
-  }),
-  locale: prefSpec<string>({
-    extract: (value) => (typeof value === "string" && isSupportedLocale(value) ? value : undefined),
-    local: (settings) => settings.locale,
-    write: (value) => ({ locale: value }),
-    clearable: true,
-    reset: () => ({ locale: undefined }),
-  }),
-  chatShowThinking: prefSpec<boolean>({
-    extract: (value) => (typeof value === "boolean" ? value : undefined),
-    local: (settings) => settings.chatShowThinking,
-  }),
-  chatShowToolCalls: prefSpec<boolean>({
-    extract: (value) => (typeof value === "boolean" ? value : undefined),
-    local: (settings) => settings.chatShowToolCalls,
-  }),
-  chatPersistCommentary: prefSpec<boolean>({
-    extract: (value) => (typeof value === "boolean" ? value : undefined),
-    local: (settings) => settings.chatPersistCommentary !== false,
-  }),
-  chatSendShortcut: prefSpec<ChatSendShortcut>({
-    extract: (value) =>
-      value === "enter" || value === "modifier-enter"
-        ? normalizeChatSendShortcut(value)
-        : undefined,
-    local: (settings) => normalizeChatSendShortcut(settings.chatSendShortcut),
-    write: (value) => ({ chatSendShortcut: value }),
-    clearable: true,
-    reset: () => ({ chatSendShortcut: undefined }),
-  }),
-  chatFollowUpMode: prefSpec<ChatFollowUpMode>({
-    extract: (value) => normalizeChatFollowUpModeOverride(value),
-    local: (settings) => normalizeChatFollowUpModeOverride(settings.chatFollowUpMode),
-    write: (value) => ({ chatFollowUpMode: value }),
-    // Unset means "use the server-configured queue mode"; clearing must propagate,
-    // so the push serializes an explicit null removal.
-    clearable: true,
-    reset: () => ({ chatFollowUpMode: undefined }),
-  }),
-  sidebarEntries: prefSpec<string[]>({
-    extract: (value) => normalizeSidebarEntries(value) ?? undefined,
-    local: (settings) => settings.sidebarEntries,
-  }),
-} as const;
-type SyncedPrefKey = keyof typeof SYNCED_PREFS;
-type ResettableServerUiPrefKey =
-  | "theme"
-  | "themeMode"
-  | "locale"
-  | "chatSendShortcut"
-  | "chatFollowUpMode";
-type SyncedPrefValue<K extends SyncedPrefKey> =
-  ReturnType<(typeof SYNCED_PREFS)[K]["extract"]> extends (infer T) | undefined ? T : never;
-type ServerUiPrefs = { [K in SyncedPrefKey]?: SyncedPrefValue<K> | null };
+  extractServerUiPrefs,
+  prefValuesEqual,
+  resolveServerUiPrefStateFromSnapshot,
+  serverPrefsLocalPatch,
+  SYNCED_PREF_KEYS,
+  SYNCED_PREFS,
+  type ResettableServerUiPrefKey,
+  type ServerUiPrefs,
+  type ServerUiPrefState,
+  type SyncedPrefKey,
+  type SyncedPrefValue,
+} from "./server-prefs-state.ts";
+import { loadSettings, patchSettings, type UiSettings } from "./settings.ts";
+import type { ThemeName } from "./theme.ts";
+
 type ServerUiPrefsWriter = Pick<RuntimeConfigCapability, "canPatch" | "runExternalMutation"> & {
   readonly state: {
     readonly client: GatewayBrowserClient | null;
@@ -122,28 +31,7 @@ type ServerUiPrefsCommit = {
   needsRefresh: boolean;
   retainedLocal?: boolean;
 };
-export type ServerUiPrefProvenance = "default" | "pending" | "synced" | "device-local";
-export type ServerUiPrefState<T> = {
-  overridden: boolean;
-  provenance: ServerUiPrefProvenance;
-  resetValue: T | undefined;
-  value: T | undefined;
-};
-const SYNCED_PREF_KEYS = Object.keys(SYNCED_PREFS) as SyncedPrefKey[];
-function extractServerUiPrefs(configObject: unknown): ServerUiPrefs {
-  const prefs = asRecord(asRecord(asRecord(configObject)?.ui)?.prefs);
-  if (!prefs) {
-    return {};
-  }
-  const result: ServerUiPrefs = {};
-  for (const key of SYNCED_PREF_KEYS) {
-    const value = SYNCED_PREFS[key].extract(prefs[key]);
-    if (value !== undefined) {
-      (result as Record<string, unknown>)[key] = value;
-    }
-  }
-  return result;
-}
+export type { ServerUiPrefProvenance, ServerUiPrefState } from "./server-prefs-state.ts";
 
 export function resolveServerUiPrefState<K extends SyncedPrefKey>(
   configObject: unknown,
@@ -152,129 +40,15 @@ export function resolveServerUiPrefState<K extends SyncedPrefKey>(
   settings = loadSettings(),
   options: { canSync?: boolean | null } = {},
 ): ServerUiPrefState<SyncedPrefValue<K>> {
-  const specification = SYNCED_PREFS[key];
-  const localValue = specification.local(settings) as SyncedPrefValue<K> | undefined;
-  const resetPatch = specification.reset?.(settings);
-  const productDefault = (
-    resetPatch ? specification.local({ ...settings, ...resetPatch }) : undefined
-  ) as SyncedPrefValue<K> | undefined;
-  const localState = (
-    resetValue: SyncedPrefValue<K> | undefined,
-  ): ServerUiPrefState<SyncedPrefValue<K>> => {
-    const overridden = !prefValuesEqual(localValue, resetValue);
-    return {
-      overridden,
-      provenance: overridden ? "device-local" : "default",
-      resetValue,
-      value: localValue,
-    };
-  };
-  const prefs = asRecord(asRecord(asRecord(configObject)?.ui)?.prefs);
-  const serverValue =
-    prefs && Object.hasOwn(prefs, key)
-      ? (specification.extract(prefs[key]) as SyncedPrefValue<K> | undefined)
-      : undefined;
-  const canApplyServerValue =
-    serverValue !== undefined &&
-    (!specification.canApply ||
-      (specification.canApply as (value: unknown, settings: UiSettings) => boolean)(
-        serverValue,
-        settings,
-      ));
-  const applicableServerValue = canApplyServerValue ? serverValue : productDefault;
   const shadowPrefs =
     scope === pendingScope ? pendingPrefs : parseStoredPrefs(readStorage(PENDING_KEY, scope));
-  if (shadowPrefs && key in shadowPrefs) {
-    if (options.canSync === false) {
-      return {
-        ...localState(applicableServerValue),
-        // The queued intent remains available for a later authorized reconnect,
-        // but this connected browser cannot claim that the value is pending sync.
-        provenance: "device-local",
-      };
-    }
-    const shadowValue = shadowPrefs[key];
-    if (shadowValue === null) {
-      return { ...localState(productDefault), provenance: "pending" };
-    }
-    return {
-      overridden: true,
-      provenance: "pending",
-      resetValue: productDefault,
-      value: shadowValue as SyncedPrefValue<K>,
-    };
-  }
-  if (!prefs || !Object.hasOwn(prefs, key)) {
-    return localState(productDefault);
-  }
-  if (serverValue === undefined) {
-    return localState(productDefault);
-  }
-  if (!canApplyServerValue) {
-    if (options.canSync === false) {
-      return localState(productDefault);
-    }
-    // The server still owns this authored preference even when this device cannot
-    // render it. Preserve that provenance so Restore default removes the override.
-    return {
-      overridden: true,
-      provenance: "synced",
-      resetValue: productDefault,
-      value: localValue,
-    };
-  }
-  if (prefValuesEqual(localValue, serverValue)) {
-    return {
-      overridden: true,
-      provenance: "synced",
-      resetValue: productDefault,
-      value: serverValue,
-    };
-  }
-  return localState(serverValue);
-}
-/** Local-settings patch that would bring the mirror in line with the server. */
-function serverPrefsLocalPatch(
-  prefs: ServerUiPrefs,
-  settings: UiSettings,
-): Partial<UiSettings> | null {
-  const patch: Partial<UiSettings> = {};
-  for (const key of SYNCED_PREF_KEYS) {
-    const specification = SYNCED_PREFS[key];
-    const serverValue = prefs[key];
-    if (serverValue === undefined) {
-      continue;
-    }
-    // Null marks a server-side removal of a clearable key: drop the local override
-    // so this device falls back to the server-configured behavior.
-    if (serverValue === null) {
-      const resetPatch = specification.clearable ? specification.reset?.(settings) : undefined;
-      if (resetPatch) {
-        for (const [resetKey, resetValue] of Object.entries(resetPatch)) {
-          if (
-            !prefValuesEqual((settings as unknown as Record<string, unknown>)[resetKey], resetValue)
-          ) {
-            (patch as Record<string, unknown>)[resetKey] = resetValue;
-          }
-        }
-      }
-      continue;
-    }
-    if (prefValuesEqual(serverValue, specification.local(settings))) {
-      continue;
-    }
-    if (
-      specification.canApply &&
-      !(specification.canApply as (value: unknown, settings: UiSettings) => boolean)(
-        serverValue,
-        settings,
-      )
-    ) {
-      continue;
-    }
-    (patch as Record<string, unknown>)[key] = serverValue;
-  }
-  return Object.keys(patch).length > 0 ? patch : null;
+  return resolveServerUiPrefStateFromSnapshot(
+    configObject,
+    key,
+    shadowPrefs,
+    settings,
+    options.canSync,
+  );
 }
 /** Synced-key delta between two local settings snapshots, for the push path. */
 export function changedServerUiPrefs(previous: UiSettings, next: UiSettings): ServerUiPrefs | null {
@@ -498,7 +272,7 @@ function reconcilePersistedPendingPrefs(): void {
     return;
   }
   const current = stored.prefs ?? {};
-  for (const key of [...pendingPersistedKeys]) {
+  for (const key of pendingPersistedKeys) {
     if (!Object.hasOwn(current, key)) {
       delete pendingPrefs[key];
       pendingPersistedKeys.delete(key);
